refactor: moved backdrop to selector for loading animation
